add seperate step to make miche tree
